+++ packages/domain-event/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.7.0
+
+### Patch Changes
+Event Searching V2
+- Updated dependencies []:
+  - @volontariapp/contracts@4.3.4
+  - @volontariapp/messaging@2.10.3
+  - @volontariapp/database@3.4.6
+  - @volontariapp/outbox@0.9.43
+
 ## 3.6.15
 
 ### Patch Changes