+++ packages/messaging/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.10.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/contracts@4.3.4
+
 ## 2.10.2
 
 ### Patch Changes